add action for after GearmanWorker work is called.
This allows for client code to do custom actions after a GearmanWorker has worked on a task.  Where this is particularly useful is if client code needs to run its own decision process for whether to kill a worker or not in the case where the plugin is set to run multiple jobs per worker.  Existing hooks in `WpGears\Gearman\Worker\do_job` cannot be used in that case because Gearman will always consider the job failed and retry.  Any exit must happen AFTER the return from GearmanWorker::work(), not while executing it.