Delay init until Plugins Loaded
I'm using WP Minions with a custom Client and Worker implementation (AWS SQS) and ran into difficulties with the order of loading. `Plugin::enable()` loads the client and worker objects immediately, but if this plugin is loaded before the plugin containing the custom classes, PHP crashes because it can't find the classes registered as the client and worker. Delaying the running of `wp_async_task_init()` until `plugins_loaded` fixes this problem.